+++ source/patterns/@aws-solutions-constructs/aws-events-rule-lambda/lib/index.ts
@@ -23,7 +23,7 @@ import { overrideProps } from '@aws-solutions-constructs/core';
  */
 export interface EventsRuleToLambdaProps {
   /**
-   * Existing instance of Lambda Function object, if this is set then the lambdaFunctionProps is ignored.
+   * Existing instance of Lambda Function object, providing both this and `lambdaFunctionProps` will cause an error.
    *
    * @default - None
    */
@@ -56,6 +56,7 @@ export class EventsRuleToLambda extends Construct {
    */
   constructor(scope: Construct, id: string, props: EventsRuleToLambdaProps) {
     super(scope, id);
+    defaults.CheckProps(props);
 
     this.lambdaFunction = defaults.buildLambdaFunction(this, {
       existingLambdaObj: props.existingLambdaObj,