+++ source/patterns/@aws-solutions-constructs/aws-lambda-s3/test/lambda-s3.test.ts
@@ -356,7 +356,7 @@ test("Test minimal deployment with an existing VPC", () => {
 });
 
 // --------------------------------------------------------------
-// Test minimal deployment with an existing VPC and existing Lambda function not in a VPCs
+// Test minimal deployment with an existing VPC and existing Lambda function not in a VPC
 //
 // buildLambdaFunction should throw an error if the Lambda function is not
 // attached to a VPC
@@ -472,4 +472,32 @@ test("Test bad call with existingBucket and bucketProps", () => {
   };
   // Assertion
   expect(app).toThrowError();
+});
+
+test('Test that CheckProps() is flagging errors correctly', () => {
+  // Stack
+  const stack = new Stack();
+
+  const testLambdaFunction = new lambda.Function(stack, 'test-lamba', {
+    runtime: lambda.Runtime.NODEJS_10_X,
+    handler: "index.handler",
+    code: lambda.Code.fromAsset(`${__dirname}/lambda`),
+  });
+
+  const app = () => {
+    new LambdaToS3(stack, "lambda-to-s3-stack", {
+      existingLambdaObj: testLambdaFunction,
+      lambdaFunctionProps: {
+        runtime: lambda.Runtime.NODEJS_10_X,
+        handler: "index.handler",
+        code: lambda.Code.fromAsset(`${__dirname}/lambda`),
+      },
+    });
+  };
+
+  // Assertion
+  expect(app).toThrowError(
+    'Error - Either provide lambdaFunctionProps or existingLambdaObj, but not both.\n'
+  );
+
 });