+++ source/patterns/@aws-solutions-constructs/aws-lambda-secretsmanager/test/lambda-secretsmanager.test.ts
@@ -77,7 +77,6 @@ test('Test deployment w/ existing secret', () => {
       handler: 'index.handler',
       code: lambda.Code.fromAsset(`${__dirname}/lambda`)
     },
-    secretProps: { removalPolicy: RemovalPolicy.DESTROY },
     existingSecretObj: existingSecret
   });
   // Assertion 1
@@ -100,10 +99,10 @@ test('Test deployment w/ existing function', () => {
     handler: 'index.handler',
     code: lambda.Code.fromAsset(`${__dirname}/lambda`)
   };
-  const existingFuntion = defaults.deployLambdaFunction(stack, lambdaFunctionProps);
+  const existingFunction = defaults.deployLambdaFunction(stack, lambdaFunctionProps);
 
   const pattern = new LambdaToSecretsmanager(stack, 'lambda-to-secretsmanager-stack', {
-    existingLambdaObj: existingFuntion,
+    existingLambdaObj: existingFunction,
     secretProps: { removalPolicy: RemovalPolicy.DESTROY },
   });
   // Assertion 1
@@ -111,7 +110,7 @@ test('Test deployment w/ existing function', () => {
     GenerateSecretString: {},
   });
   // Assertion 2
-  expect(pattern.lambdaFunction).toBe(existingFuntion);
+  expect(pattern.lambdaFunction).toBe(existingFunction);
 });
 
 // --------------------------------------------------------------
@@ -292,7 +291,7 @@ test("Test minimal deployment with an existing VPC", () => {
 });
 
 // --------------------------------------------------------------
-// Test minimal deployment with an existing VPC and existing Lambda function not in a VPCs
+// Test minimal deployment with an existing VPC and existing Lambda function not in a VPC
 //
 // buildLambdaFunction should throw an error if the Lambda function is not
 // attached to a VPC